Extract new window policy handler to a new file.
